requirements.txt
metadrive-simulator
stable-baselines3[extra]
torch
pyyaml
timm
gymnasium